Features/assets (#6)
* Added support for assets.

Created the new assets resource. This resource is used to add files and images to the Contentful environment.

* Fixed versioning bug.

When creating and updating assets and entries, the publish and unpublish calls also updates the version property. This resulted in a version mismatch error.

* Fixed early-return bug.

Setting properties call seems to throw an error when called with Docker. Removed return should fix this problem.

Co-authored-by: Kantoor <office@bithopper.io>